🔧 Update: Add payment verification fields to Appointment model, including *payment_method*, *payment_verified_at*, *payment_verified_by*, *payment_notes*, *receipt_hash*, and *receipt_url*. Update payment handling in AppointmentViewSet to save verification details. Update payment amount handling in various components to use *parseFloat* instead of *parseInt*. Add *isValidEndTime* helper function to validate end time against start time.

+++ royal-care-frontend/src/components/OperatorDashboard.jsx
@@ -1195,7 +1195,7 @@ const OperatorDashboard = () => {
     });
 
     // Validate payment data
-    if (!paymentData.amount || parseInt(paymentData.amount) <= 0) {
+    if (!paymentData.amount || parseFloat(paymentData.amount) <= 0) {
       console.log(
         "❌ handleMarkPaymentPaid: Invalid payment amount",
         paymentData.amount
@@ -1225,7 +1225,7 @@ const OperatorDashboard = () => {
       // Ensure payment amount is properly formatted as a number
       const processedPaymentData = {
         ...paymentData,
-        amount: parseInt(paymentData.amount) || 0, // Ensure it's an integer
+        amount: parseFloat(paymentData.amount) || 0, // Ensure it's a number
       };
 
       console.log(
